+++ dotenv/mod_test.ts
@@ -6,6 +6,7 @@ import {
   assertStringIncludes,
   assertThrows,
 } from "../testing/asserts.js";
+import type { DotenvConfig } from "./mod.js";
 import {
   config,
   configSync,
@@ -14,6 +15,8 @@ import {
   stringify,
 } from "./mod.js";
 import * as path from "../path/mod.js";
+import type { IsExact } from "../testing/types.js";
+import { assertType } from "../testing/types.js";
 
 const moduleDir = path.dirname(path.fromFileUrl(import.meta.url));
 const testdataDir = path.resolve(moduleDir, "testdata");
@@ -711,4 +714,48 @@ Deno.test("use of restrictEnvAccessTo for an Env var, without granting env permi
   const error = decoder.decode(stdout).trim();
 
   assertStringIncludes(error, 'Requires env access to "GREETING"');
+});
+
+Deno.test("type inference based on restrictEnvAccessTo", async (t) => {
+  await t.step("return type is inferred", async () => {
+    const conf = await config({
+      ...testOptions,
+      restrictEnvAccessTo: ["GREETING"],
+    });
+
+    assertType<
+      IsExact<typeof conf, { GREETING: string }>
+    >(true);
+
+    assertType<
+      IsExact<typeof conf, { NO_SUCH_KEY: string }>
+    >(false);
+
+    assertType<
+      IsExact<typeof conf, DotenvConfig>
+    >(false);
+
+    assertEquals(conf.DEFAULT1, "Some Default");
+  });
+
+  await t.step("readonly array is also supported", () => {
+    const conf = configSync({
+      ...testOptions,
+      restrictEnvAccessTo: ["GREETING", "DEFAULT1"] as const,
+    });
+
+    assertType<
+      IsExact<typeof conf, { GREETING: string; DEFAULT1: string }>
+    >(true);
+  });
+
+  await t.step("without restrictEnvAccessTo", async () => {
+    const conf = await config(testOptions);
+
+    assertType<
+      IsExact<typeof conf, { GREETING: string }>
+    >(false);
+
+    assertType<IsExact<typeof conf, DotenvConfig>>(true);
+  });
 });