+++ io/buffer.ts
@@ -1,7 +1,8 @@
 // Copyright 2018-2022 the Deno authors. All rights reserved. MIT license.
 import { assert } from "../_util/asserts.ts";
 import { BytesList } from "../bytes/bytes_list.ts";
-import { concat, copy } from "../bytes/mod.ts";
+import { concat } from "../bytes/concat.ts";
+import { copy } from "../bytes/copy.ts";
 import type { Reader, ReaderSync, Writer, WriterSync } from "./types.d.ts";
 
 // MIN_READ is the minimum ArrayBuffer size passed to a read call by
@@ -41,7 +42,7 @@ export class Buffer {
    * `reset()`, or `truncate()`). If `options.copy` is false the slice aliases the buffer content at
    * least until the next buffer modification, so immediate changes to the
    * slice will affect the result of future reads.
-   * @param options Defaults to `{ copy: true }`
+   * @param [options={ copy: true }]
    */
   bytes(options = { copy: true }): Uint8Array {
     if (options.copy === false) return this.#buf.subarray(this.#off);
@@ -968,7 +969,22 @@ export async function* readDelim(
   }
 }
 
-/** Read delimited strings from a Reader. */
+/**
+ * Read Reader chunk by chunk, splitting based on delimiter.
+ *
+ * @example
+ * ```ts
+ * import { readStringDelim } from "https://deno.land/std@$STD_VERSION/io/mod.ts";
+ * import * as path from "https://deno.land/std@$STD_VERSION/path/mod.ts";
+ *
+ * const filename = path.join(Deno.cwd(), "std/io/README.md");
+ * let fileReader = await Deno.open(filename);
+ *
+ * for await (let line of readStringDelim(fileReader, "\n")) {
+ *   console.log(line);
+ * }
+ * ```
+ */
 export async function* readStringDelim(
   reader: Reader,
   delim: string,
@@ -985,7 +1001,22 @@ export async function* readStringDelim(
   }
 }
 
-/** Read strings line-by-line from a Reader. */
+/**
+ * Read strings line-by-line from a Reader.
+ *
+ *  @example
+ * ```ts
+ * import { readLines } from "https://deno.land/std@$STD_VERSION/io/mod.ts";
+ * import * as path from "https://deno.land/std@$STD_VERSION/path/mod.ts";
+ *
+ * const filename = path.join(Deno.cwd(), "std/io/README.md");
+ * let fileReader = await Deno.open(filename);
+ *
+ * for await (let line of readLines(fileReader)) {
+ *   console.log(line);
+ * }
+ * ```
+ */
 export async function* readLines(
   reader: Reader,
   decoderOpts?: {