+++ node/fs.ts
@@ -53,12 +53,13 @@ import {
   writeFileSync,
 } from "./_fs/_fs_writeFile.js";
 import { Stats } from "./internal/fs/utils.mjs";
+// @deno-types="./internal/fs/streams.d.ts"
 import {
   createReadStream,
   createWriteStream,
   ReadStream,
   WriteStream,
-} from "./internal/fs/streams.js";
+} from "./internal/fs/streams.mjs";
 
 const {
   F_OK,