+++ node/http.ts
@@ -1,7 +1,6 @@
 // Copyright 2018-2022 the Deno authors. All rights reserved. MIT license.
 
 import { type Deferred, deferred } from "../async/deferred.js";
-import { core } from "./_core.js";
 import { _normalizeArgs, ListenOptions, Socket } from "./net.js";
 import { Buffer } from "./buffer.js";
 import { ERR_SERVER_NOT_RUNNING } from "./internal/errors.js";
@@ -15,6 +14,7 @@ import {
 } from "./stream.js";
 import { OutgoingMessage } from "./_http_outgoing.js";
 import { Agent } from "./_http_agent.mjs";
+import { chunkExpression as RE_TE_CHUNKED } from "./_http_common.js";
 import { urlToHttpOptions } from "./internal/url.js";
 import { constants, TCP } from "./internal_binding/tcp_wrap.js";
 
@@ -63,9 +63,10 @@ const DenoServe = Deno[Deno.internal]?.nodeUnstable?.serve || Deno.serve;
 const DenoUpgradeHttpRaw = Deno[Deno.internal]?.nodeUnstable?.upgradeHttpRaw ||
   Deno.upgradeHttpRaw;
 
+const ENCODER = new TextEncoder();
 function chunkToU8(chunk: Chunk): Uint8Array {
   if (typeof chunk === "string") {
-    return core.encode(chunk);
+    return ENCODER.encode(chunk);
   }
   return chunk;
 }
@@ -96,6 +97,7 @@ export interface RequestOptions {
   href?: string;
 }
 
+// TODO: Implement ClientRequest methods (e.g. setHeader())
 /** ClientRequest represents the http(s) request from the client */
 class ClientRequest extends NodeWritable {
   defaultProtocol = "http:";
@@ -130,8 +132,14 @@ class ClientRequest extends NodeWritable {
       this.controller.close();
     }
 
+    const body = await this._createBody(this.body, this.opts);
     const client = await this._createCustomClient();
-    const opts = { body: this.body, method: this.opts.method, client };
+    const opts = {
+      body,
+      method: this.opts.method,
+      client,
+      headers: this.opts.headers,
+    };
     const mayResponse = fetch(this._createUrlStrFromOptions(this.opts), opts)
       .catch((e) => {
         if (e.message.includes("connection closed before message completed")) {
@@ -156,6 +164,31 @@ class ClientRequest extends NodeWritable {
 
   abort() {
     this.destroy();
+  }
+
+  async _createBody(
+    body: ReadableStream | null,
+    opts: RequestOptions,
+  ): Promise<Buffer | ReadableStream | null> {
+    if (!body) return null;
+    if (!opts.headers) return body;
+
+    const headers = Object.fromEntries(
+      Object.entries(opts.headers).map(([k, v]) => [k.toLowerCase(), v]),
+    );
+
+    if (
+      !RE_TE_CHUNKED.test(headers["transfer-encoding"]) &&
+      !Number.isNaN(Number.parseInt(headers["content-length"], 10))
+    ) {
+      const bufferList: Buffer[] = [];
+      for await (const chunk of body) {
+        bufferList.push(chunk);
+      }
+      return Buffer.concat(bufferList);
+    }
+
+    return body;
   }
 
   _createCustomClient(): Promise<Deno.HttpClient | undefined> {
@@ -195,6 +228,7 @@ class ClientRequest extends NodeWritable {
 /** IncomingMessage for http(s) client */
 export class IncomingMessageForClient extends NodeReadable {
   reader: ReadableStreamDefaultReader | undefined;
+  #statusMessage = "";
   constructor(public response: Response | undefined, public socket: Socket) {
     super();
     this.reader = response?.body?.getReader();
@@ -234,7 +268,11 @@ export class IncomingMessageForClient extends NodeReadable {
   }
 
   get statusMessage() {
-    return this.response?.statusText || "";
+    return this.#statusMessage || this.response?.statusText || "";
+  }
+
+  set statusMessage(v: string) {
+    this.#statusMessage = v;
   }
 }
 
@@ -419,7 +457,7 @@ export class IncomingMessageForServer extends NodeReadable {
     });
     // TODO: consider more robust path extraction, e.g:
     // url: (new URL(request.url).pathname),
-    this.url = req.url.slice(req.url.indexOf("/", 8));
+    this.url = req.url?.slice(req.url.indexOf("/", 8));
     this.method = req.method;
     this.#req = req;
   }