+++ node/http_test.ts
@@ -1,12 +1,13 @@
 // Copyright 2018-2022 the Deno authors. All rights reserved. MIT license.
 
 import EventEmitter from "./events.js";
-import http from "./http.js";
+import http, { type RequestOptions } from "./http.js";
 import { ERR_SERVER_NOT_RUNNING } from "./internal/errors.js";
 import { assert, assertEquals } from "../testing/asserts.js";
 import { deferred } from "../async/deferred.js";
 import { gzip } from "./zlib.js";
 import { Buffer } from "./buffer.js";
+import { serve } from "../http/server.js";
 
 Deno.test("[node/http listen]", async () => {
   {
@@ -151,6 +152,35 @@ Deno.test("[node/http] request default protocol", async () => {
   await promise;
 });
 
+Deno.test("[node/http] request with headers", async () => {
+  const promise = deferred<void>();
+  const server = http.createServer((req, res) => {
+    assertEquals(req.headers["x-foo"], "bar");
+    res.end("ok");
+  });
+  server.listen(() => {
+    const req = http.request(
+      {
+        host: "localhost",
+        port: server.address().port,
+        headers: { "x-foo": "bar" },
+      },
+      (res) => {
+        res.on("data", () => {});
+        res.on("end", () => {
+          server.close();
+        });
+        assertEquals(res.statusCode, 200);
+      },
+    );
+    req.end();
+  });
+  server.on("close", () => {
+    promise.resolve();
+  });
+  await promise;
+});
+
 Deno.test("[node/http] non-string buffer response", async () => {
   const promise = deferred<void>();
   const server = http.createServer((_, res) => {
@@ -178,4 +208,188 @@ Deno.test("[node/http] non-string buffer response", async () => {
     promise.resolve();
   });
   await promise;
+});
+
+Deno.test("[node/http] http.IncomingMessage can be created without url", () => {
+  const message = new http.IncomingMessage(
+    // adapted from https://github.com/dougmoscrop/serverless-http/blob/80bfb3e940057d694874a8b0bc12ad96d2abe7ab/lib/request.js#L7
+    {
+      // @ts-expect-error - non-request properties will also be passed in, e.g. by serverless-http
+      encrypted: true,
+      readable: false,
+      remoteAddress: "foo",
+      address: () => ({ port: 443 }),
+      end: Function.prototype,
+      destroy: Function.prototype,
+    },
+  );
+  message.url = "https://example.com";
+});
+
+Deno.test("[node/http] set http.IncomingMessage.statusMessage", () => {
+  const message = new http.IncomingMessageForClient(
+    new Response(null, { status: 404, statusText: "Not Found" }),
+    {
+      encrypted: true,
+      readable: false,
+      remoteAddress: "foo",
+      // @ts-expect-error - good enough for this test
+      address() {
+        return { port: 443, family: "IPv4" };
+      },
+      end(_cb) {
+        return this;
+      },
+      destroy(_e) {
+        return;
+      },
+    },
+  );
+  assertEquals(message.statusMessage, "Not Found");
+  message.statusMessage = "boom";
+  assertEquals(message.statusMessage, "boom");
+});
+
+Deno.test("[node/http] send request with non-chunked body", async () => {
+  let requestHeaders: Headers;
+  let requestBody = "";
+
+  const hostname = "localhost";
+  const port = 4505;
+
+  // NOTE: Instead of node/http.createServer(), serve() in std/http/server.ts is used.
+  // https://github.com/denoland/deno_std/pull/2755#discussion_r1005592634
+  const handler = async (req: Request) => {
+    requestHeaders = req.headers;
+    requestBody = await req.text();
+    return new Response("ok");
+  };
+  const abortController = new AbortController();
+  const servePromise = serve(handler, {
+    hostname,
+    port,
+    signal: abortController.signal,
+    onListen: undefined,
+  });
+
+  const opts: RequestOptions = {
+    host: hostname,
+    port,
+    method: "POST",
+    headers: {
+      "Content-Type": "text/plain; charset=utf-8",
+      "Content-Length": "11",
+    },
+  };
+  const req = http.request(opts, (res) => {
+    res.on("data", () => {});
+    res.on("end", () => {
+      abortController.abort();
+    });
+    assertEquals(res.statusCode, 200);
+    assertEquals(requestHeaders.get("content-length"), "11");
+    assertEquals(requestHeaders.has("transfer-encoding"), false);
+    assertEquals(requestBody, "hello world");
+  });
+  req.write("hello ");
+  req.write("world");
+  req.end();
+
+  await servePromise;
+});
+
+Deno.test("[node/http] send request with chunked body", async () => {
+  let requestHeaders: Headers;
+  let requestBody = "";
+
+  const hostname = "localhost";
+  const port = 4505;
+
+  // NOTE: Instead of node/http.createServer(), serve() in std/http/server.ts is used.
+  // https://github.com/denoland/deno_std/pull/2755#discussion_r1005592634
+  const handler = async (req: Request) => {
+    requestHeaders = req.headers;
+    requestBody = await req.text();
+    return new Response("ok");
+  };
+  const abortController = new AbortController();
+  const servePromise = serve(handler, {
+    hostname,
+    port,
+    signal: abortController.signal,
+    onListen: undefined,
+  });
+
+  const opts: RequestOptions = {
+    host: hostname,
+    port,
+    method: "POST",
+    headers: {
+      "Content-Type": "text/plain; charset=utf-8",
+      "Content-Length": "11",
+      "Transfer-Encoding": "chunked",
+    },
+  };
+  const req = http.request(opts, (res) => {
+    res.on("data", () => {});
+    res.on("end", () => {
+      abortController.abort();
+    });
+    assertEquals(res.statusCode, 200);
+    assertEquals(requestHeaders.has("content-length"), false);
+    assertEquals(requestHeaders.get("transfer-encoding"), "chunked");
+    assertEquals(requestBody, "hello world");
+  });
+  req.write("hello ");
+  req.write("world");
+  req.end();
+
+  await servePromise;
+});
+
+Deno.test("[node/http] send request with chunked body as default", async () => {
+  let requestHeaders: Headers;
+  let requestBody = "";
+
+  const hostname = "localhost";
+  const port = 4505;
+
+  // NOTE: Instead of node/http.createServer(), serve() in std/http/server.ts is used.
+  // https://github.com/denoland/deno_std/pull/2755#discussion_r1005592634
+  const handler = async (req: Request) => {
+    requestHeaders = req.headers;
+    requestBody = await req.text();
+    return new Response("ok");
+  };
+  const abortController = new AbortController();
+  const servePromise = serve(handler, {
+    hostname,
+    port,
+    signal: abortController.signal,
+    onListen: undefined,
+  });
+
+  const opts: RequestOptions = {
+    host: hostname,
+    port,
+    method: "POST",
+    headers: {
+      "Content-Type": "text/plain; charset=utf-8",
+    },
+  };
+  const req = http.request(opts, (res) => {
+    res.on("data", () => {});
+    res.on("end", () => {
+      abortController.abort();
+    });
+    assertEquals(res.statusCode, 200);
+    assertEquals(requestHeaders.has("content-length"), false);
+    assertEquals(requestHeaders.get("transfer-encoding"), "chunked");
+    assertEquals(requestBody, "hello world");
+  });
+  req.write("hello ");
+  req.write("world");
+  req.end();
+
+  await servePromise;
 });